+ Finished first release with full game logic/scoring.

+++ CardGames/Basas/BasasConsts.cs
@@ -10,6 +10,6 @@
         /// <summary>
         /// Number of cards per each player in each round.
         /// </summary>
-        public static readonly int[] RoundCards = new int[] { 7, 6, 5, 4, 3, 2, 1, 1, 2, 3, 4, 5, 6, 7 };
+        public static readonly byte[] RoundCards = new byte[] { 7, 6, 5, 4, 3, 2, 1, 1, 2, 3, 4, 5, 6, 7 };
     }
 }
